Struct cpython::oldstyle::PyClass

pub struct PyClass<'p>(_);

Represents an old-style Python class.

Only available with Python 2.x.

Methods

impl<'p> PyClass<'p>

fn is_subclass_of(&self, base: &PyClass<'p>) -> bool

Return true if self is a subclass of base.

fn create_instance<T>(&self, args: T, kw: Option<&PyDict<'p>>) -> PyResult<'p, PyInstance<'p>> where T: ToPyObject<'p, ObjectType=PyTuple<'p>>

Create a new instance of the class. The parameters args and kw are used as the positional and keyword parameters to the object’s constructor.

fn create_instance_raw(&self, dict: PyDict<'p>) -> PyResult<'p, PyInstance<'p>>

Create a new instance of a specific class without calling its constructor. The dict parameter will be used as the object’s dict.