Module cpython::py_class

Modules

gc

Enums

CompareOp

Traits

BaseObject
A PythonObject that is usable as a base type with the py_class!() macro.

PythonObjectFromPyClassMacro
Trait implemented by the types produced by the py_class!() macro.