Trait cpython::py_class::PythonObjectFromPyClassMacro

pub trait PythonObjectFromPyClassMacro: PythonObjectWithTypeObject {
    fn initialize(py: Python, module_name: Option<&str>) -> PyResult<PyType>;

    fn add_to_module(py: Python, module: &PyModule) -> PyResult<()>;
}

Trait implemented by the types produced by the py_class!() macro.

This is an unstable implementation detail; do not implement manually!

Required Methods

fn initialize(py: Python, module_name: Option<&str>) -> PyResult<PyType>

Initializes the class.

module_name: the name of the parent module into which the class will be placed.

fn add_to_module(py: Python, module: &PyModule) -> PyResult<()>

Initializes the class and adds it to the module.