Struct cpython::GILProtected

pub struct GILProtected<T> {
    // some fields omitted
}

Mutex-like wrapper object for data that is protected by the Python GIL.

Example

use std::cell::Cell;
use cpython::{Python, GILProtected};

let data = GILProtected::new(Cell::new(0));

{
    let gil_guard = Python::acquire_gil();
    let cell = data.get(gil_guard.python());
    cell.set(cell.get() + 1);
}

Methods

impl<T> GILProtected<T>

fn new(data: T) -> GILProtected<T>

Creates a new instance of GILProtected.

fn get<'a>(&'a self, _py: Python<'a>) -> &'a T

Returns a shared reference to the data stored in the GILProtected.

Requires a Python instance as proof that the GIL is acquired.

fn into_inner(self) -> T

Consumes the GILProtected, returning the wrapped value.