Struct cpython::PyBytes

pub struct PyBytes(_);

Represents a Python byte string. Corresponds to str in Python 2, and bytes in Python 3.

Methods

impl PyBytes

fn new(py: Python, s: &[u8]) -> PyBytes

Creates a new Python byte string object. The byte string is initialized by copying the data from the &[u8].

Panics if out of memory.

fn as_slice(&self, _py: Python) -> &[u8]

Gets the Python string data as byte slice.

impl PyString

fn extract<'a>(py: Python, o: &'a PyObject) -> PyResult<Cow<'a, str>>

Extract a rust string from the Python object.

In Python 2.7, accepts both byte strings and unicode strings. Byte strings will be decoded using UTF-8.

In Python 3.x, accepts unicode strings only.

Returns TypeError if the input is not one of the accepted types. Returns UnicodeDecodeError if the input is not valid unicode.

fn extract_lossy<'a>(py: Python, o: &'a PyObject) -> PyResult<Cow<'a, str>>

Extract a rust string from the Python object.

In Python 2.7, accepts both byte strings and unicode strings. Byte strings will be decoded using UTF-8.

In Python 3.x, accepts unicode strings only.

Returns TypeError if the input is not one of the accepted types. Any invalid code points are replaced with U+FFFD REPLACEMENT CHARACTER.

fn to_string(&self, py: Python) -> PyResult<Cow<str>>

Convert the PyString into a rust string.

In Python 2.7, PyString is a byte string and will be decoded using UTF-8. In Python 3.x, PyString is a unicode string.

Returns a UnicodeDecodeError if the input is not valid unicode.

fn to_string_lossy(&self, py: Python) -> Cow<str>

Convert the PyString into a rust string.

In Python 2.7, PyString is a byte string and will be decoded using UTF-8. In Python 3.x, PyString is a unicode string.

Any invalid UTF-8 sequences are replaced with U+FFFD REPLACEMENT CHARACTER.