Struct cpython::PyDict

pub struct PyDict<'p>(_);

Represents a Python dict.

Methods

impl<'p> PyDict<'p>

fn new(py: Python<'p>) -> PyDict<'p>

Creates a new empty dictionary.

May panic when running out of memory.

fn copy(&self) -> PyResult<'p, PyDict<'p>>

Return a new dictionary that contains the same key-value pairs as self. Corresponds to dict(self) in Python.

fn clear(&self)

Empty an existing dictionary of all key-value pairs.

fn len(&self) -> usize

Return the number of items in the dictionary. This is equivalent to len(p) on a dictionary.

fn contains<K>(&self, key: K) -> PyResult<'p, bool> where K: ToPyObject<'p>

Determine if the dictionary contains the specified key. This is equivalent to the Python expression key in self.

fn get_item<K>(&self, key: K) -> Option<PyObject<'p>> where K: ToPyObject<'p>

Gets an item from the dictionary. Returns None if the item is not present, or if an error occurs.

fn set_item<K, V>(&self, key: K, value: V) -> PyResult<'p, ()> where K: ToPyObject<'p>, V: ToPyObject<'p>

Sets an item value. This is equivalent to the Python expression self[key] = value.

fn del_item<K>(&self, key: K) -> PyResult<'p, ()> where K: ToPyObject<'p>

Deletes an item. This is equivalent to the Python expression del self[key].

fn items_list(&self) -> PyList<'p>

fn items(&self) -> Vec<(PyObject<'p>, PyObject<'p>)>

Returns the list of (key,value) pairs in this dictionary.