Struct cpython::PyInt

pub struct PyInt<'p>(_);

Represents a Python int object.

Note that in Python 2.x, int and long are different types. When rust-cpython is compiled for Python 3.x, PyInt and PyLong are aliases for the same type, which corresponds to a Python int.

You can usually avoid directly working with this type by using ToPyObject and extract with the primitive Rust integer types.

Methods

impl<'p> PyInt<'p>

fn new(py: Python<'p>, val: c_long) -> PyInt<'p>

Creates a new Python int object.

fn value(&self) -> c_long

Gets the value of this integer.