Struct cpython::PyLong

pub struct PyLong<'p>(_);

In Python 2.x, represents a Python long object. In Python 3.x, represents a Python int object. Both PyInt and PyLong refer to the same type on Python 3.x.

You can usually avoid directly working with this type by using ToPyObject and extract with the primitive Rust integer types.