Struct cpython::PyModule

pub struct PyModule(_);

Represents a Python module object.

Implementations

impl PyModule

pub fn new(py: Python<'_>, name: &str) -> PyResult<PyModule>

Create a new module object with the __name__ attribute set to name.

pub fn import(py: Python<'_>, name: &str) -> PyResult<PyModule>

Import the Python module with the specified name.

pub fn dict(&self, py: Python<'_>) -> PyDict

Return the dictionary object that implements module's namespace; this object is the same as the __dict__ attribute of the module object.

pub fn name<'a>(&'a self, py: Python<'_>) -> PyResult<&'a str>

Gets the module name.

May fail if the module does not have a __name__ attribute.

pub fn filename(&self, py: Python<'_>) -> PyResult<&str>

Gets the module filename.

May fail if the module does not have a __file__ attribute.

pub fn get(&self, py: Python<'_>, name: &str) -> PyResult<PyObject>

Gets a member from the module. This is equivalent to the Python expression: getattr(module, name)

pub fn call<A>(
&self,
py: Python<'_>,
name: &str,
args: A,
kwargs: Option<&PyDict>
) -> PyResult<PyObject> where
A: ToPyObject<ObjectType = PyTuple>,

Calls a function in the module. This is equivalent to the Python expression: getattr(module, name)(*args, **kwargs)

args should be a value that, when converted to Python, results in a tuple. For this purpose, you can use:

cpython::NoArgs when calling a method without any arguments

otherwise, a Rust tuple with 1 or more elements

Example

use cpython::NoArgs;
let sys = py.import("sys").unwrap();
// Call function without arguments:
let encoding = sys.call(py, "getdefaultencoding", NoArgs, None).unwrap();
// Call function with a single argument:
sys.call(py, "setrecursionlimit", (1000,), None).unwrap();

pub fn add<V>(&self, py: Python<'_>, name: &str, value: V) -> PyResult<()> where
V: ToPyObject,

Adds a member to the module.

This is a convenience function which can be used from the module's initialization function.

pub fn add_class<T>(&self, py: Python<'_>) -> PyResult<()> where
T: PythonObjectFromPyClassMacro,

Adds a new extension type to the module.

This is a convenience function that initializes the py_class!(), sets new_type.__module__ to this module's name, and adds the type to this module.