Struct cpython::PyRustObject

pub struct PyRustObject<T, B = PyObject> where T: 'static + Send, B: PythonBaseObject {
    // some fields omitted
}

A Python object that contains a rust value of type T, and is derived from base class B. Note that this type effectively acts like Rc<T>, except that the reference counting is done by the Python runtime.

Methods

impl<T, B> PyRustObject<T, B> where T: 'static + Send, B: PythonBaseObject

fn as_base(&self) -> &B

Gets a reference to this object, but of the base class type.

fn into_base(self) -> B

Gets a reference to this object, but of the base class type.

fn get<'a>(&'a self, _py: Python<'a>) -> &'a T

Gets a reference to the rust value stored in this Python object.