Struct cpython::PyRustType

pub struct PyRustType<'p, T, B = PyObject<'p>> where T: 'p + Send, B: PythonBaseObject<'p> {
    // some fields omitted
}

A Python class that contains rust values of type T. Serves as a Python type object, and can be used to construct PyRustObject<T> instances.

Methods

impl<'p, T, B> PyRustType<'p, T, B> where T: 'p + Send, B: PythonBaseObject<'p>

fn create_instance(&self, val: T, base_val: B::InitType) -> PyRustObject<'p, T, B>

Creates a PyRustObject instance from a value.