Struct cpython::PyRustTypeBuilder

#[must_use]
pub struct PyRustTypeBuilder<'p, T, B = PyObject<'p>> where T: 'static + Send, B: PythonBaseObject<'p> {
    // some fields omitted
}

Methods

impl<'p, T> PyRustTypeBuilder<'p, T> where T: 'static + Send

fn new(py: Python<'p>, name: &str) -> PyRustTypeBuilder<'p, T>

Create a new type builder.

py: proof that the GIL is held by the current thread. name: name of the new type

fn base<T2, B2>(self, base_type: &PyRustType<'p, T2, B2>) -> PyRustTypeBuilder<'p, T, PyRustObject<'p, T2, B2>> where T2: 'static + Send, B2: PythonBaseObject<'p>

Sets the base class that this type is inheriting from.

impl<'p, T, B> PyRustTypeBuilder<'p, T, B> where T: 'static + Send, B: PythonBaseObject<'p>

fn doc(self, doc_str: &str) -> Self

Set the doc string on the type being built.

fn add<M>(self, name: &str, val: M) -> Self where M: TypeMember<'p, PyRustObject<'p, T, B>>

Adds a new member to the type.

fn finish(self) -> PyResult<'p, PyRustType<'p, T, B>>

Finalize construction of the new type.