Struct cpython::PySequence

pub struct PySequence(/* private fields */);

Represents a reference to a python object supporting the sequence protocol.

Implementations

impl PySequence

pub fn len(&self, py: Python<'_>) -> PyResult<isize>

Returns the number of objects in sequence. This is equivalent to Python len().

pub fn concat(&self, py: Python<'_>, other: &PySequence) -> PyResult<PyObject>

Return the concatenation of o1 and o2. Equivalent to python o1 + o2

pub fn repeat(&self, py: Python<'_>, count: isize) -> PyResult<PyObject>

Return the result of repeating sequence object o count times. Equivalent to python o * count NB: Python accepts negative counts; it returns an empty Sequence.

pub fn in_place_concat( &self, py: Python<'_>, other: &PySequence ) -> PyResult<PyObject>

Return the concatenation of o1 and o2 on success. Equivalent to python o1 += o2

pub fn in_place_repeat( &self, py: Python<'_>, count: isize ) -> PyResult<PyObject>

Return the result of repeating sequence object o count times. Equivalent to python o *= count NB: Python accepts negative counts; it empties the Sequence.

pub fn get_item(&self, py: Python<'_>, index: isize) -> PyResult<PyObject>

Return the ith element of the Sequence. Equivalent to python o[index]

pub fn get_slice( &self, py: Python<'_>, begin: isize, end: isize ) -> PyResult<PyObject>

Return the slice of sequence object o between begin and end. This is the equivalent of the Python expression o[begin:end]

pub fn set_item(&self, py: Python<'_>, i: isize, v: &PyObject) -> PyResult<()>

Assign object v to the ith element of o. Equivalent to Python statement o[i] = v

pub fn del_item(&self, py: Python<'_>, i: isize) -> PyResult<()>

Delete the ith element of object o. Python statement del o[i]

pub fn set_slice( &self, py: Python<'_>, i1: isize, i2: isize, v: &PyObject ) -> PyResult<()>

Assign the sequence object v to the slice in sequence object o from i1 to i2. This is the equivalent of the Python statement o[i1:i2] = v

pub fn del_slice(&self, py: Python<'_>, i1: isize, i2: isize) -> PyResult<()>

Delete the slice in sequence object o from i1 to i2. equivalent of the Python statement del o[i1:i2]

pub fn count<V>(&self, py: Python<'_>, value: V) -> PyResult<usize>
where V: ToPyObject,

Return the number of occurrences of value in o, that is, return the number of keys for which o[key] == value

pub fn contains<V>(&self, py: Python<'_>, value: V) -> PyResult<bool>
where V: ToPyObject,

Determine if o contains value. this is equivalent to the Python expression value in o

pub fn index<V>(&self, py: Python<'_>, value: V) -> PyResult<usize>
where V: ToPyObject,

Return the first index i for which o[i] == value. This is equivalent to the Python expression o.index(value)

pub fn list(&self, py: Python<'_>) -> PyResult<PyList>

Return a fresh list based on the Sequence.

pub fn tuple(&self, py: Python<'_>) -> PyResult<PyTuple>

Return a fresh tuple based on the Sequence.

pub fn iter<'p>(&self, py: Python<'p>) -> PyResult<PyIterator<'p>>