Struct cpython::PySharedRef

pub struct PySharedRef<'a, T: 'a + ?Sized> { /* fields omitted */ }

A reference to PySharedRefCell owned by a Python object.

This is a lifetime-bound reference to the PySharedRefCell data field, and will be created by the automatically generated accessor function.

impl MyType {
    fn data_name<'a>(&'a self, py: Python<'a>) -> PySharedRef<'a, T> { ... }
}

Implementations

impl<'a, T: ?Sized> PySharedRef<'a, T>

pub fn borrow(&self) -> Ref<'a, T>

Immutably borrows the wrapped value.

Panics

Panics if the value is currently mutably borrowed.

pub fn try_borrow(&self) -> Result<Ref<'a, T>, BorrowError>

Immutably borrows the wrapped value, returning an error if the value is currently mutably borrowed.

pub fn borrow_mut(&self) -> RefMut<'a, T>

Mutably borrows the wrapped value.

Any existing leaked references will be invalidated.

Panics

Panics if the value is currently borrowed.

pub fn try_borrow_mut(&self) -> Result<RefMut<'a, T>, BorrowMutError>

Mutably borrows the wrapped value, returning an error if the value is currently borrowed.

pub fn leak_immutable(&self) -> UnsafePyLeaked<&'static T>

Creates an immutable reference which is not bound to lifetime.

Panics

Panics if the value is currently mutably borrowed.

pub fn try_leak_immutable(
&self
) -> Result<UnsafePyLeaked<&'static T>, BorrowError>

Creates an immutable reference which is not bound to lifetime, returning an error if the value is currently mutably borrowed.