Crate cpython

Rust bindings to the python interpreter.

Ownership and Lifetimes

In python, all objects are implicitly reference counted. In rust, we will use the PyObject type to represent a reference to a python object.

Because all python objects potentially have multiple owners, the concept concept of rust mutability does not apply to python objects. As a result, this API will allow mutating python objects even if they are not stored in a mutable rust variable.

The python interpreter uses a global interpreter lock (GIL) to ensure thread-safety. This API uses the lifetime parameter PyObject<'p> to ensure that python objects cannot be accessed without holding the GIL. Throughout this library, the lifetime 'p always refers to the lifetime of the python interpreter.

When accessing existing objects, the lifetime on PyObject<'p> is sufficient to ensure that the GIL is held by the current code. But we also need to ensure that the GIL is held when creating new objects. For this purpose, this library uses the marker type Python<'p>, which acts like a reference to the whole python interpreter.

You can obtain a Python<'p> instance by acquiring the GIL, or by calling python() on any existing python object.

Error Handling

The vast majority of operations in this library will return PyResult<'p, ...>. This is an alias for the type Result<..., PyErr<'p>>.

A PyErr represents a python exception. Errors within the rust-cpython library are also exposed as python exceptions.

Example

extern crate cpython;

use cpython::{PythonObject, Python};
 
fn main() {
    let gil_guard = Python::acquire_gil();
    let py = gil_guard.python();
    let sys = py.import("sys").unwrap();
    let version = sys.get("version").unwrap().extract::<String>().unwrap();
    println!("Hello Python {}", version);
}

Modules

exc
This module contains the python exception types.

oldstyle
This module contains support for old-style classes. Only available in Python 2.x.

Macros

py_func!
Creates a python callable object that invokes a Rust function.

py_module_initializer!
Expands to an extern "C" function that allows python to load the rust code as a python extension module.

Structs

GILGuard
RAII type that represents an acquired GIL.

PyBool

PyDict

PyErr
Represents a python exception that was raised.

PyFloat

PyInt

PyIterator

PyList

PyLong

PyModule

PyObject

PyString

PyTuple

PyType

PyUnicode

Python
Marker type that indicates that the GIL is currently held.

Traits

FromPyObject
FromPyObject is implemented by various types that can be extracted from a python object.

ObjectProtocol
Trait that contains methods

PythonObject
Trait implemented by all python object types.

PythonObjectWithCheckedDowncast
Trait implemented by python object types that allow a checked downcast.

PythonObjectWithTypeObject
Trait implemented by python object types that have a corresponding type object.

ToPyObject
Conversion trait that allows various objects to be converted into python objects.

ToPythonPointer
This trait allows retrieving the underlying FFI pointer from python objects.

Functions

prepare_freethreaded_python
Prepares the use of python in a free-threaded context.

Type Definitions

PyResult
Represents the result of a python call.

Py_ssize_t