Trait cpython::buffer::Element

pub unsafe trait Element {
    fn is_compatible_format(format: &CStr) -> bool;
}

Trait implemented for possible element types of PyBuffer.

Required Methods

fn is_compatible_format(format: &CStr) -> bool

Gets whether the element specified in the format string is potentially compatible. Alignment and size are checked separately from this function.