Struct cpython::PyDict

pub struct PyDict(_);

Represents a Python dict.

Methods

impl PyDict

fn new(py: Python) -> PyDict

Creates a new empty dictionary.

May panic when running out of memory.

fn copy(&self, py: Python) -> PyResult<PyDict>

Return a new dictionary that contains the same key-value pairs as self. Corresponds to dict(self) in Python.

fn clear(&self, _py: Python)

Empty an existing dictionary of all key-value pairs.

fn len(&self, _py: Python) -> usize

Return the number of items in the dictionary. This is equivalent to len(p) on a dictionary.

fn contains<K>(&self, py: Python, key: K) -> PyResult<bool> where K: ToPyObject

Determine if the dictionary contains the specified key. This is equivalent to the Python expression key in self.

fn get_item<K>(&self, py: Python, key: K) -> Option<PyObject> where K: ToPyObject

Gets an item from the dictionary. Returns None if the item is not present, or if an error occurs.

fn set_item<K, V>(&self, py: Python, key: K, value: V) -> PyResult<()> where K: ToPyObject, V: ToPyObject

Sets an item value. This is equivalent to the Python expression self[key] = value.

fn del_item<K>(&self, py: Python, key: K) -> PyResult<()> where K: ToPyObject

Deletes an item. This is equivalent to the Python expression del self[key].

fn items_list(&self, py: Python) -> PyList

fn items(&self, py: Python) -> Vec<(PyObject, PyObject)>

Returns the list of (key,value) pairs in this dictionary.