Struct cpython::PyErr

pub struct PyErr<'p> {
    pub ptype: PyObject<'p>,
    pub pvalue: Option<PyObject<'p>>,
    pub ptraceback: Option<PyObject<'p>>,
}

Represents a Python exception that was raised.

Fields

ptype

The type of the exception. This should be either a PyClass or a PyType.

pvalue

The value of the exception.

This can be either an instance of ptype, a tuple of arguments to be passed to ptype's constructor, or a single argument to be passed to ptype's constructor. Call PyErr::instance() to get the exception instance in all cases.

ptraceback

The PyTraceBack object associated with the error.

Methods

impl<'p> PyErr<'p>

fn occurred(_: Python<'p>) -> bool

Gets whether an error is present in the Python interpreter's global state.

fn fetch(py: Python<'p>) -> PyErr<'p>

Retrieves the current error from the Python interpreter's global state. The error is cleared from the Python interpreter. If no error is set, returns a SystemError.

fn new<O>(obj: O) -> PyErr<'p> where O: PythonObject<'p>

Creates a new PyErr.

If obj is a Python exception instance, the PyErr will use that instance. If obj is a Python exception type object, the PyErr will (lazily) create a new instance of that type. Otherwise, a TypeError is created instead.

fn new_lazy_init(exc: PyType<'p>, value: Option<PyObject<'p>>) -> PyErr<'p>

Construct a new error, with the usual lazy initialization of Python exceptions. exc is the exception type; usually one of the standard exceptions like PyExc::runtime_error(). value is the exception instance, or a tuple of arguments to pass to the exception constructor.

fn print(self)

Print a standard traceback to sys.stderr.

fn print_and_set_sys_last_vars(self)

Print a standard traceback to sys.stderr.

fn matches(&self, exc: &PyObject) -> bool

Return true if the current exception matches the exception in exc. If exc is a class object, this also returns true when self is an instance of a subclass. If exc is a tuple, all exceptions in the tuple (and recursively in subtuples) are searched for a match.

fn normalize(&mut self)

Normalizes the error. This ensures that the exception value is an instance of the exception type.

fn get_type(&self) -> PyType<'p>

Retrieves the exception type.

If the exception type is an old-style class, returns oldstyle::PyClass.

fn instance(&mut self) -> PyObject<'p>

Retrieves the exception instance for this error. This method takes &mut self because the error might need to be normalized in order to create the exception instance.

fn restore(self)

Writes the error back to the Python interpreter's global state. This is the opposite of PyErr::fetch().

fn warn(py: Python<'p>, category: &PyObject, message: &str, stacklevel: i32) -> PyResult<'p, ()>

Issue a warning message. May return a PyErr if warnings-as-errors is enabled.