Struct cpython::PyModule

pub struct PyModule<'p>(_);

Represents a Python module object.

Methods

impl<'p> PyModule<'p>

fn new(py: Python<'p>, name: &str) -> PyResult<'p, PyModule<'p>>

Create a new module object with the __name__ attribute set to name.

fn import(py: Python<'p>, name: &str) -> PyResult<'p, PyModule<'p>>

Import the Python module with the specified name.

fn dict(&self) -> PyDict<'p>

Return the dictionary object that implements module's namespace; this object is the same as the __dict__ attribute of the module object.

fn name<'a>(&'a self) -> PyResult<'p, &'a str>

Gets the module name.

May fail if the module does not have a __name__ attribute.

fn filename<'a>(&'a self) -> PyResult<'p, &'a str>

Gets the module filename.

May fail if the module does not have a __file__ attribute.

fn get(&self, name: &str) -> PyResult<'p, PyObject<'p>>

Gets a member from the module. This is equivalent to the Python expression: getattr(module, name)

fn call<A>(&self, name: &str, args: A, kwargs: Option<&PyDict<'p>>) -> PyResult<'p, PyObject<'p>> where A: ToPyObject<'p, ObjectType=PyTuple<'p>>

Calls a function in the module. This is equivalent to the Python expression: getattr(module, name)(*args, **kwargs)

fn add<V>(&self, name: &str, value: V) -> PyResult<'p, ()> where V: ToPyObject<'p>

Adds a member to the module.

This is a convenience function which can be used from the module's initialization function.

fn add_type<T>(&self, name: &str) -> PyRustTypeBuilder<'p, T> where T: 'p + Send

Adds a new extension type to the module.

This is a convenience function that creates a new PyRustTypeBuilder and sets new_type.__module__ to this module's name. The new type will be added to this module when finish() is called on the builder.