Struct cpython::PyModule

pub struct PyModule(_);

Represents a Python module object.

Methods

impl PyModule

fn new(py: Python, name: &str) -> PyResult<PyModule>

Create a new module object with the __name__ attribute set to name.

fn import(py: Python, name: &str) -> PyResult<PyModule>

Import the Python module with the specified name.

fn dict(&self, py: Python) -> PyDict

Return the dictionary object that implements module's namespace; this object is the same as the __dict__ attribute of the module object.

fn name<'a>(&'a self, py: Python) -> PyResult<&'a str>

Gets the module name.

May fail if the module does not have a __name__ attribute.

fn filename<'a>(&'a self, py: Python) -> PyResult<&'a str>

Gets the module filename.

May fail if the module does not have a __file__ attribute.

fn get(&self, py: Python, name: &str) -> PyResult<PyObject>

Gets a member from the module. This is equivalent to the Python expression: getattr(module, name)

fn call<A>(&self, py: Python, name: &str, args: A, kwargs: Option<&PyDict>) -> PyResult<PyObject> where A: ToPyObject<ObjectType=PyTuple>

Calls a function in the module. This is equivalent to the Python expression: getattr(module, name)(*args, **kwargs)

fn add<V>(&self, py: Python, name: &str, value: V) -> PyResult<()> where V: ToPyObject

Adds a member to the module.

This is a convenience function which can be used from the module's initialization function.

fn add_type<'p, T>(&self, py: Python<'p>, name: &str) -> PyRustTypeBuilder<'p, T> where T: 'static + Send

Adds a new extension type to the module.

This is a convenience function that creates a new PyRustTypeBuilder and sets new_type.__module__ to this module's name. The new type will be added to this module when finish() is called on the builder.