Struct cpython::PyObject

#[repr(C)]

pub struct PyObject { /* fields omitted */ }

Represents a reference to a Python object.

Python objects are reference counted. Calling clone_ref() on a PyObject will return a new reference to the same object (thus incrementing the reference count). The Drop implementation will automatically decrement the reference count. You can also call release_ref() to explicitly decrement the reference count. This is slightly faster than relying on automatic drop, because release_ref does not need to check whether the GIL needs to be acquired.

PyObject can be used with all Python objects, since all python types derive from object. This crate also contains other, more specific types that serve as references to Python objects (e.g. PyTuple for Python tuples, etc.).

You can convert from any Python object to PyObject by calling as_object() or into_object() from the PythonObject trait. In the other direction, you can call cast_as() or cast_into() on PyObject to convert to more specific object types.

Most of the interesting methods are provided by the ObjectProtocol trait.

Methods

impl PyObject

pub unsafe fn from_owned_ptr(_py: Python, ptr: *mut PyObject) -> PyObject

Creates a PyObject instance for the given FFI pointer. This moves ownership over the pointer into the PyObject. Undefined behavior if the pointer is NULL or invalid.

pub unsafe fn from_borrowed_ptr(_py: Python, ptr: *mut PyObject) -> PyObject

Creates a PyObject instance for the given FFI pointer. Calls Py_INCREF() on the ptr. Undefined behavior if the pointer is NULL or invalid.

pub unsafe fn from_owned_ptr_opt(
py: Python,
ptr: *mut PyObject
) -> Option<PyObject>

Creates a PyObject instance for the given FFI pointer. This moves ownership over the pointer into the PyObject. Returns None for null pointers; undefined behavior if the pointer is invalid.

pub unsafe fn from_borrowed_ptr_opt(
py: Python,
ptr: *mut PyObject
) -> Option<PyObject>

Returns None for null pointers; undefined behavior if the pointer is invalid.

pub fn as_ptr(&self) -> *mut PyObject

Gets the underlying FFI pointer. Returns a borrowed pointer.

pub fn steal_ptr(self) -> *mut PyObject

Gets the underlying FFI pointer. Consumes self without calling Py_DECREF(), thus returning an owned pointer.

pub unsafe fn borrow_from_ptr<'a>(ptr: &'a *mut PyObject) -> &'a PyObject

Transmutes an FFI pointer to &PyObject. Undefined behavior if the pointer is NULL or invalid.

pub unsafe fn borrow_from_owned_ptr_slice<'a>(
ptr: &'a [*mut PyObject]
) -> &'a [PyObject]

Transmutes a slice of owned FFI pointers to &[PyObject]. Undefined behavior if any pointer in the slice is NULL or invalid.

pub fn get_refcnt(&self, _py: Python) -> usize

Gets the reference count of this Python object.

pub fn get_type(&self, py: Python) -> PyType

Gets the Python type object for this object's type.

pub unsafe fn unchecked_cast_into<T>(self) -> T where
T: PythonObject,

Casts the PyObject to a concrete Python object type. Causes undefined behavior if the object is not of the expected type. This is a wrapper function around PythonObject::unchecked_downcast_from().

pub fn cast_into<'p, T>(
self,
py: Python<'p>
) -> Result<T, PythonObjectDowncastError<'p>> where
T: PythonObjectWithCheckedDowncast,

Casts the PyObject to a concrete Python object type. Fails with PythonObjectDowncastError if the object is not of the expected type. This is a wrapper function around PythonObjectWithCheckedDowncast::downcast_from().

Important traits for &'a mut I
impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub unsafe fn unchecked_cast_as<'s, T>(&'s self) -> &'s T where
T: PythonObject,

Casts the PyObject to a concrete Python object type. Causes undefined behavior if the object is not of the expected type. This is a wrapper function around PythonObject::unchecked_downcast_borrow_from().

pub fn cast_as<'s, 'p, T>(
&'s self,
py: Python<'p>
) -> Result<&'s T, PythonObjectDowncastError<'p>> where
T: PythonObjectWithCheckedDowncast,

Casts the PyObject to a concrete Python object type. Fails with PythonObjectDowncastError if the object is not of the expected type. This is a wrapper function around PythonObjectWithCheckedDowncast::downcast_borrow_from().

pub fn extract<'a, T>(&'a self, py: Python) -> PyResult<T> where
T: FromPyObject<'a>,

Extracts some type from the Python object. This is a wrapper function around FromPyObject::from_py_object().