Struct cpython::PyObject

pub struct PyObject<'p> {
    // some fields omitted
}

Represents a reference to a Python object.

Python objects are reference counted. Calling clone() on a PyObject will return a new reference to the same object (thus incrementing the reference count). The Drop implementation will decrement the reference count.

PyObject can be used with all Python objects, since all python types derive from object. This crate also contains other, more specific types that serve as references to Python objects (e.g. PyTuple for Python tuples, etc.).

You can convert from any Python object to PyObject by calling as_object() or into_object from the PythonObject trait. In the other direction, you can call cast_as or cast_into on PyObject to convert to more specific object types.

Most of the interesting methods are provided by the ObjectProtocol trait.

Methods

impl<'p> PyObject<'p>

unsafe fn from_owned_ptr(py: Python<'p>, ptr: *mut PyObject) -> PyObject<'p>

Creates a PyObject instance for the given FFI pointer. This moves ownership over the pointer into the PyObject. Undefined behavior if the pointer is NULL or invalid.

unsafe fn from_borrowed_ptr(py: Python<'p>, ptr: *mut PyObject) -> PyObject<'p>

Creates a PyObject instance for the given FFI pointer. Calls Py_INCREF() on the ptr. Undefined behavior if the pointer is NULL or invalid.

unsafe fn from_owned_ptr_opt(py: Python<'p>, ptr: *mut PyObject) -> Option<PyObject<'p>>

Creates a PyObject instance for the given FFI pointer. This moves ownership over the pointer into the PyObject. Returns None for null pointers; undefined behavior if the pointer is invalid.

unsafe fn from_borrowed_ptr_opt(py: Python<'p>, ptr: *mut PyObject) -> Option<PyObject<'p>>

Returns None for null pointers; undefined behavior if the pointer is invalid.

fn as_ptr(&self) -> *mut PyObject

Gets the underlying FFI pointer. Returns a borrowed pointer.

fn steal_ptr(self) -> *mut PyObject

Gets the underlying FFI pointer. Consumes self without calling Py_DECREF(), thus returning an owned pointer.

unsafe fn borrow_from_owned_ptr<'a>(py: Python<'p>, ptr: &'a *mut PyObject) -> &'a PyObject<'p>

Transmutes an owned FFI pointer to &PyObject. Undefined behavior if the pointer is NULL or invalid.

unsafe fn borrow_from_owned_ptr_slice<'a>(py: Python<'p>, ptr: &'a [*mut PyObject]) -> &'a [PyObject<'p>]

Transmutes a slice of owned FFI pointers to &[PyObject]. Undefined behavior if the pointer is NULL or invalid.

fn get_refcnt(&self) -> usize

Gets the reference count of this Python object.

fn get_type(&self) -> &PyType<'p>

Gets the Python type object for this object's type.

unsafe fn unchecked_cast_into<T>(self) -> T where T: PythonObject<'p>

Casts the PyObject to a concrete Python object type. Causes undefined behavior if the object is not of the expected type. This is a wrapper function around PythonObject::unchecked_downcast_from().

fn cast_into<T>(self) -> Result<T, PythonObjectDowncastError<'p>> where T: PythonObjectWithCheckedDowncast<'p>

Casts the PyObject to a concrete Python object type. Fails with PythonObjectDowncastError if the object is not of the expected type. This is a wrapper function around PythonObjectWithCheckedDowncast::downcast_from().

unsafe fn unchecked_cast_as<'s, T>(&'s self) -> &'s T where T: PythonObject<'p>

Casts the PyObject to a concrete Python object type. Causes undefined behavior if the object is not of the expected type. This is a wrapper function around PythonObject::unchecked_downcast_borrow_from().

fn cast_as<'s, T>(&'s self) -> Result<&'s T, PythonObjectDowncastError<'p>> where T: PythonObjectWithCheckedDowncast<'p>

Casts the PyObject to a concrete Python object type. Fails with PythonObjectDowncastError if the object is not of the expected type. This is a wrapper function around PythonObjectWithCheckedDowncast::downcast_borrow_from().

fn extract<T>(&self) -> Result<T, PyErr<'p>> where T: FromPyObject<'p>

Extracts some type from the Python object. This is a wrapper function around FromPyObject::from_py_object().