Struct cpython::PyRustObject

pub struct PyRustObject<'p, T, B = PyObject<'p>> where T: 'static, B: PythonBaseObject<'p> {
    // some fields omitted
}

A Python object that contains a rust value of type T, and is derived from base class B. Note that this type effectively acts like Rc<T>, except that the reference counting is done by the Python runtime.

Methods

impl<'p, T, B> PyRustObject<'p, T, B> where T: 'static + Send, B: PythonBaseObject<'p>

fn base(&self) -> &B

Gets a reference to this object, but of the base class type.

fn get(&self) -> &T

Gets a reference to the rust value stored in this Python object.