Struct cpython::PyType

pub struct PyType(_);

Represents a reference to a Python type object.

Methods

impl PyType

fn as_type_ptr(&self) -> *mut PyTypeObject

Retrieves the underlying FFI pointer associated with this Python object.

unsafe fn from_type_ptr(py: Python, p: *mut PyTypeObject) -> PyType

Retrieves the PyType instance for the given FFI pointer. This increments the reference count on the type object. Undefined behavior if the pointer is NULL or invalid.

fn is_subtype_of(&self, _: Python, b: &PyType) -> bool

Return true if self is a subtype of b.

fn is_instance(&self, _: Python, obj: &PyObject) -> bool

Return true if obj is an instance of self.

fn call<A>(&self, py: Python, args: A, kwargs: Option<&PyDict>) -> PyResult<PyObject> where A: ToPyObject<ObjectType=PyTuple>

Calls the type object, thus creating a new instance. This is equivalent to the Python expression: self(*args, **kwargs)