Struct cpython::PyType

pub struct PyType<'p>(_);

Represents a reference to a Python type object.

Methods

impl<'p> PyType<'p>

fn as_type_ptr(&self) -> *mut PyTypeObject

Retrieves the underlying FFI pointer associated with this Python object.

unsafe fn from_type_ptr<'a>(py: Python<'p>, p: *mut PyTypeObject) -> PyType<'p>

Retrieves the PyType instance for the given FFI pointer. Undefined behavior if the pointer is NULL or invalid.

fn is_subtype_of(&self, b: &PyType<'p>) -> bool

Return true if self is a subtype of b.

fn is_instance(&self, obj: &PyObject<'p>) -> bool

Return true if obj is an instance of self.

fn call<A>(&self, args: A, kwargs: Option<&PyDict<'p>>) -> PyResult<'p, PyObject<'p>> where A: ToPyObject<'p, ObjectType=PyTuple<'p>>

Calls the type object, thus creating a new instance. This is equivalent to the Python expression: self(*args, **kwargs)