Struct cpython::Python

pub struct Python<'p>(_);

Marker type that indicates that the GIL is currently held.

The 'Python' struct is a zero-size marker struct that is required for most Python operations. This is used to indicate that the operation accesses/modifies the Python interpreter state, and thus can only be called if the Python interpreter is initialized and the Python global interpreter lock (GIL) is acquired. The lifetime 'p represents the lifetime of the Python interpreter.

You can imagine the GIL to be a giant Mutex<PythonInterpreterState>. The type Python<'p> then acts like a reference &'p PythonInterpreterState.

Methods

impl<'p> Python<'p>

unsafe fn assume_gil_acquired() -> Python<'p>

Retrieve Python instance under the assumption that the GIL is already acquired at this point, and stays acquired for the lifetime 'p.

Because the output lifetime 'p is not connected to any input parameter, care must be taken that the compiler infers an appropriate lifetime for 'p when calling this function.

fn acquire_gil() -> GILGuard

Acquires the global interpreter lock, which allows access to the Python runtime.

If the Python runtime is not already initialized, this function will initialize it. See prepare_freethreaded_python() for details.

unsafe fn allow_threads<T, F>(self, f: F) -> T where F: FnOnce() -> T

Releases the GIL and allows the use of Python on other threads. Unsafe because we do not ensure that existing references to Python objects are not accessed within the closure.

fn None(self) -> PyObject<'p>

Gets the Python builtin value None.

fn True(self) -> PyBool<'p>

Gets the Python builtin value True.

fn False(self) -> PyBool<'p>

Gets the Python builtin value False.

fn get_type<T>(self) -> PyType<'p> where T: PythonObjectWithTypeObject<'p>

Gets the Python type object for type T.

fn import(self, name: &str) -> PyResult<'p, PyModule<'p>>

Import the Python module with the specified name.