Trait cpython::FromPyObject

pub trait FromPyObject<'s>: Sized {
    fn extract(py: Python<'_>, obj: &'s PyObject) -> PyResult<Self>;
}

FromPyObject is implemented by various types that can be extracted from a Python object.

Normal usage is through the PyObject::extract helper method:

let value = obj.extract::<TargetType>(py)?;

Each target type for this conversion supports a different Python objects as input. Calls with an unsupported input object will result in an exception (usually a TypeError).

This trait is also used by the py_fn! and py_class! and py_argparse! macros in order to translate from Python objects to the expected Rust parameter types. For example, the parameter x in def method(self, x: i32) will use impl FromPyObject for i32 to convert the input Python object into a Rust i32. When these macros are used with reference parameters (x: &str), the trait RefFromPyObject is used instead.

Required methods

fn extract(py: Python<'_>, obj: &'s PyObject) -> PyResult<Self>

Extracts Self from the source PyObject.

Loading content...