Trait cpython::PythonObject

pub trait PythonObject<'p>: 'p + Clone {
    fn as_object(&self) -> &PyObject<'p>;
    fn into_object(self) -> PyObject<'p>;
    unsafe fn unchecked_downcast_from(PyObject<'p>) -> Self;
    unsafe fn unchecked_downcast_borrow_from<'a>(&'a PyObject<'p>) -> &'a Self;

    fn python(&self) -> Python<'p> { ... }
}

Trait implemented by all Python object types.

Required Methods

fn as_object(&self) -> &PyObject<'p>

Casts the Python object to PyObject.

fn into_object(self) -> PyObject<'p>

Casts the Python object to PyObject.

unsafe fn unchecked_downcast_from(PyObject<'p>) -> Self

Unchecked downcast from PyObject to Self. Undefined behavior if the input object does not have the expected type.

unsafe fn unchecked_downcast_borrow_from<'a>(&'a PyObject<'p>) -> &'a Self

Unchecked downcast from PyObject to Self. Undefined behavior if the input object does not have the expected type.

Provided Methods

fn python(&self) -> Python<'p>

Retrieve Python instance from an existing Python object.