Trait cpython::PythonObjectWithTypeObject

pub trait PythonObjectWithTypeObject: PythonObjectWithCheckedDowncast {
    fn type_object(py: Python) -> PyType;
}

Trait implemented by Python object types that have a corresponding type object.

Required methods

fn type_object(py: Python) -> PyType

Retrieves the type object for this Python object type.

Loading content...