Trait cpython::PythonObjectWithTypeObject

pub trait PythonObjectWithTypeObject<'p>: PythonObjectWithCheckedDowncast<'p> {
    fn type_object(Python<'p>) -> PyType<'p>;
}

Trait implemented by Python object types that have a corresponding type object.

Required Methods

fn type_object(Python<'p>) -> PyType<'p>

Retrieves the type object for this Python object type.