Trait cpython::PythonObjectWithTypeObject

pub trait PythonObjectWithTypeObject: PythonObjectWithCheckedDowncast {
    fn type_object(Python) -> PyType;
}

Trait implemented by Python object types that have a corresponding type object.

Required Methods

fn type_object(Python) -> PyType

Retrieves the type object for this Python object type.