Trait cpython::ToPyObject

pub trait ToPyObject {
    type ObjectType: PythonObject;
    fn to_py_object(&self, py: Python) -> Self::ObjectType;

    fn into_py_object(self, py: Python) -> Self::ObjectType
    where
        Self: Sized,
    { ... }

    fn with_borrowed_ptr<F, R>(&self, py: Python, f: F) -> R
    where
        F: FnOnce(*mut PyObject) -> R,
    { ... }
}

Conversion trait that allows various objects to be converted into Python objects.

Note: The associated type ObjectType is used so that some Rust types convert to a more precise type of Python object. For example, [T]::to_py_object() will result in a PyList. You can always calls val.to_py_object(py).into_py_object() in order to obtain PyObject (the second into_py_object() call via the PythonObject trait corresponds to the upcast from PyList to PyObject).

Associated Types

type ObjectType: PythonObject

Loading content...
Required methods

fn to_py_object(&self, py: Python) -> Self::ObjectType

Converts self into a Python object.

Loading content...
Provided methods

fn into_py_object(self, py: Python) -> Self::ObjectType where
Self: Sized,

Converts self into a Python object.

May be more efficient than to_py_object in some cases because it can move out of the input object.

fn with_borrowed_ptr<F, R>(&self, py: Python, f: F) -> R where
F: FnOnce(*mut PyObject) -> R,

Converts self into a Python object and calls the specified closure on the native FFI pointer underlying the Python object.

May be more efficient than to_py_object because it does not need to touch any reference counts when the input object already is a Python object.

Loading content...